+++ packages/camera/camera_android_camerax/pigeons/camerax_library.dart
@@ -46,6 +46,46 @@ class CameraPermissionsErrorData {
   String description;
 }
 
+/// The states the camera can be in.
+///
+/// See https://developer.android.com/reference/androidx/camera/core/CameraState.Type.
+enum CameraStateType {
+  closed,
+  closing,
+  open,
+  opening,
+  pendingOpen,
+}
+
+class CameraStateTypeData {
+  late CameraStateType value;
+}
+
+/// The types (T) properly wrapped to be used as a LiveData<T>.
+///
+/// If you need to add another type to support a type S to use a LiveData<S> in
+/// this plugin, ensure the following is done on the Dart side:
+///
+///  * In `../lib/src/live_data.dart`, add new cases for S in
+///    `_LiveDataHostApiImpl#getValueFromInstances` to get the current value of
+///    type S from a LiveData<S> instance and in `LiveDataFlutterApiImpl#create`
+///    to create the expected type of LiveData<S> when requested.
+///
+/// On the native side, ensure the following is done:
+///
+///  * Update `LiveDataHostApiImpl#getValue` is updated to properly return
+///    identifiers for instances of type S.
+///  * Update `ObserverFlutterApiWrapper#onChanged` to properly handle receiving
+///    calls with instances of type S if a LiveData<S> instance is observed.
+enum LiveDataSupportedType {
+  cameraState,
+  zoomState,
+}
+
+class LiveDataSupportedTypeData {
+  late LiveDataSupportedType value;
+}
+
 class ExposureCompensationRange {
   ExposureCompensationRange({
     required this.minCompensation,
@@ -77,6 +117,8 @@ abstract class JavaObjectFlutterApi {
 @HostApi(dartHostTestHandler: 'TestCameraInfoHostApi')
 abstract class CameraInfoHostApi {
   int getSensorRotationDegrees(int identifier);
+
+  int getCameraState(int identifier);
 
   int getExposureState(int identifier);
 
@@ -228,6 +270,11 @@ abstract class ImageCaptureHostApi {
 }
 
 @FlutterApi()
+abstract class CameraStateFlutterApi {
+  void create(int identifier, CameraStateTypeData type, int? errorIdentifier);
+}
+
+@FlutterApi()
 abstract class ExposureStateFlutterApi {
   void create(
       int identifier,
@@ -252,6 +299,35 @@ abstract class ImageAnalysisHostApi {
 @HostApi(dartHostTestHandler: 'TestAnalyzerHostApi')
 abstract class AnalyzerHostApi {
   void create(int identifier);
+}
+
+@HostApi(dartHostTestHandler: 'TestObserverHostApi')
+abstract class ObserverHostApi {
+  void create(int identifier);
+}
+
+@FlutterApi()
+abstract class ObserverFlutterApi {
+  void onChanged(int identifier, int valueIdentifier);
+}
+
+@FlutterApi()
+abstract class CameraStateErrorFlutterApi {
+  void create(int identifier, int code);
+}
+
+@HostApi(dartHostTestHandler: 'TestLiveDataHostApi')
+abstract class LiveDataHostApi {
+  void observe(int identifier, int observerIdentifier);
+
+  void removeObservers(int identifier);
+
+  int? getValue(int identifier, LiveDataSupportedTypeData type);
+}
+
+@FlutterApi()
+abstract class LiveDataFlutterApi {
+  void create(int identifier, LiveDataSupportedTypeData type);
 }
 
 @FlutterApi()