+++ packages/camera/camera_android_camerax/CHANGELOG.md
@@ -23,3 +23,5 @@
 * Implements retrieval of camera information.
 * Updates README.md with plugin overview and adds contribution guide to CONTRIBUTING.md.
 * Implements video capture.
+* Implements onCameraClosing callback method for indicating the camera is closing and bumps CameraX version to 1.3.0-alpha05.
+